Move actionsCellStyle to div container
The `actionCellStyle` css style object is currently applied to the `TableCell` and not the child `<div>` element. Because of this, the styles do not allow formatting of the element such as `justifyContent: 'center'`  This simple edit moves the styles to the div. Hopefully this is the intended behavior. See (Issue #888)[https://github.com/mbrn/material-table/issues/888]